@equest/rlparser/parser/requirements.txt
numpy==1.18.2
pandas==1.0.3
setuptools==41.6.0
xlrd==1.1.0
google-api-python-client
protobuf==3.6.1
carball
matplotlib
boto3